requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.4
django-crispy-forms==2.4
idna==3.10
Markdown==3.8.2
markdownify==1.2.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.5
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
